// ev-backoffice/client/src/app/core/models/email-template-types.enum.ts
export enum EmailTemplateTypes {
  NEW_CLIENT = 'NEW_CLIENT',
  UPDATED_CLIENT = 'UPDATED_CLIENT',
  COVER_LETTER_NEW = 'COVER_LETTER_NEW',
  COVER_LETTER_UPDTAED = 'COVER_LETTER_UPDTAED',
  LEGAL_FEES_NEW = 'LEGAL_FEES_NEW',
  LEGAL_FEES_UPDATED = 'LEGAL_FEES_UPDATED',
  RETAINER_AGREEMENT_NEW = 'RETAINER_AGREEMENT_NEW',
  RETAINER_AGREEMENT_UPDATED = 'RETAINER_AGREEMENT_UPDATED',
  CLOSING_TEXT = 'CLOSING_TEXT',
  DOCUMENT_REJECTION_NOTIFICATION = 'DOCUMENT_REJECTION_NOTIFICATION',
  INVITE_COLLEAGUE_TO_EASYVISA = 'INVITE_COLLEAGUE_TO_EASYVISA',
  INVITE_APPLICANT = 'INVITE_APPLICANT',
  NEW_EMPLOYEE_REGISTRATION_INVITE = 'NEW_EMPLOYEE_REGISTRATION_INVITE',
  PACKAGE_TRANSFER_SUCCESSFUL = 'PACKAGE_TRANSFER_SUCCESSFUL',
  LIST_OF_BENEFICIARIES = 'LIST_OF_BENEFICIARIES',
  LEGAL_REPRESENTATIVE_ADDRESS = 'LEGAL_REPRESENTATIVE_ADDRESS',
  ADDITIONAL_FEES = 'ADDITIONAL_FEES',
  QUESTIONNAIRE_INACTIVITY = 'QUESTIONNAIRE_INACTIVITY',
  DOCUMENT_PORTAL_INACTIVITY = 'DOCUMENT_PORTAL_INACTIVITY',
  USCIS_FORM_EDITION_UPDATE_TO_APPLICANTS = 'USCIS_FORM_EDITION_UPDATE_TO_APPLICANTS',
  USCIS_FORM_EDITION_UPDATE_TO_ORG_MEMBERS = 'USCIS_FORM_EDITION_UPDATE_TO_ORG_MEMBERS',
  PAYMENT = 'PAYMENT',
}